services/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.37
psycopg[binary]==3.2.5
pydantic==2.10.5
pydantic-settings==2.7.1
python-multipart==0.0.20
PyYAML==6.0.2
Pillow==11.1.0
redis==5.2.1
celery==5.4.0
python-dotenv==1.0.1
orjson==3.10.15
